apps/backend/pyproject.toml
[tool.poetry]
name = "wc26-backend"
version = "0.1.0"
description = "FastAPI backend for WC26 Fantasy Friends"
authors = ["WC26 Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = "^2.0.30"
alembic = "^1.13.2"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pydantic = "^2.7.3"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
openai = "^1.30.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

